+++ components/cards/ConversationPremium/ConversationPremium.module.css
@@ -256,7 +256,7 @@
 }
 .new_name{
     border: 1px solid var(--newViolet);
-    background-color: #a361c557;
+    background-color:  var(--inputBckgr);
     margin-left: 3%;
     border-radius: 8px;
     width: 90%;
@@ -288,7 +288,7 @@
 
 .add_new_message_cont>div{
     /* background-color: var(--rosita1); */
-    background-color: #a361c557;
+    background-color:  var(--inputBckgr);
     border: 1px solid var(--newViolet);
     border-radius: 500px;
     width: 100%;
@@ -314,7 +314,7 @@
     backdrop-filter: blur(20px);
 }
 .noPrevChat span:hover{
-    background-color: #a361c557;
+    background-color:  var(--inputBckgr);
 
 }
 .noPrevChat span:first-child{
@@ -439,7 +439,7 @@
     opacity: 1;
 }
 /* split modal */
-.excludeModal{
+.splitModal{
     border: 2px solid var(--newViolet);
     border-radius: 15px;
     padding: 30px 30px 48px;
@@ -450,5 +450,33 @@
     width: 100%;
     height: 100%;
     z-index: 4;
-    background-color: rgba(0, 0, 0, 0.8);
+    /* background-color: rgba(0, 0, 0, 0.8); */
+    background-color: #17132bcc;
+    backdrop-filter: brightness(0.3);
+}
+.modalSplitCont{
+
+}
+.splitNameCont{
+    border: 1px solid var(--newViolet);
+    border-radius: 8px;
+    display: flex;
+}
+.splitNameCont span{
+    font-size: 0.7em;
+    padding: 10px 18px;
+    display: inline-block;
+    text-align: center;
+}
+.splitNameCont input{
+    font-size: 0.7em;
+    padding: 0 18px;
+    display: inline-block;
+    width: 100%;
+    border: none;
+    background-color: var(--inputBckgr);
+    outline: none;
+}
+.splitNameCont input::placeholder{
+    color: rgb(160, 160, 160);
 }
