+++ styles/globals.css
@@ -36,6 +36,7 @@ li{ list-style: none; }
 	--newPink: #917fc9;
   --newPink2: #ad93f0;
   --borders: #4d318c;
+  --inputBckgr : #a361c557;
 }
 /* 
 ::-webkit-scrollbar{width:9px;position:absolute; background-color: rgb(242, 242, 242); height: 4px;}
